[MISC] put seqan3::detail::advanceable_alignment_coordinate into own header
